Move FakeClient to chain_client and get_block_numbers to BlockingProviderExt
