+++ rust/provider/Cargo.toml
@@ -7,12 +7,14 @@ edition = "2021"
 alloy-chains = { workspace = true }
 alloy-primitives = { workspace = true, features = ["rlp", "serde"] }
 anyhow = { workspace = true }
+async-trait = { workspace = true }
 auto_impl = { workspace = true }
 block_header = { workspace = true }
 common = { workspace = true }
 derive-new = { workspace = true }
 ethers-core = { workspace = true }
 ethers-providers = { workspace = true }
+futures = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 thiserror = { workspace = true }
