--- rust/services/call/host/test_data/two_mainnet_blocks_rpc_cache.json
@@ -1,70 +1,0 @@
-{
-  "partial_blocks": [
-    [
-      {
-        "block_no": {
-          "Number": 21515063
-        }
-      },
-      {
-        "Eth": {
-          "parent_hash": "0x6e2b0f597255188947e5bf06e974ce35437d684a871481975c01cab3dd616129",
-          "ommers_hash": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347",
-          "beneficiary": "0x0000000000000000000000000000000000000000",
-          "state_root": "0x19b534bdbcc69c52953e3e4e8582f106187bfe1fba7793567fdbc60402c24046",
-          "transactions_root": "0xea1240a739420f39bb7272c1a8de298fa8d820ac4f02e7b7426264f2d0eaab0f",
-          "receipts_root": "0xde1e634f5a9ef6d892bf17aa0f4f836d1532d99369b56239ceeee09a1a2569d4",
-          "logs_bloom": "0x95ab41faedcf093cd224a3fcc170512a91d9409c6da069b0013fa81d733b91877449e541b1dca2b7aed977b0784e0d7cfecfc965acf7ffb32df71fc8c8bee1ebda234ede1e03feffeb66784adbc43ceb70ccb9f3ede43df0dd43d707f5f4df3fdf1ae00e8ff6b46f278e9e9d386e68fb00a6d4f15d958e09c22214dfc16f78050bfd9f32a139f28c7f4d94648b6ec0f64540a3edd7edcce9796bf1f6f6b276bffb673572bd9a7c107bc153d2acaf9c3035ed7751f195b313ace3ba2fc5a56db4dadeb42f9b4e0bd89948813a4779d8c5f3ec253e94ef4cfe64b6e8aaee94666243b3ff49a9b4802c2d149285a55560068644ddd14b7e32c5d5c3b8031e2b34c5",
-          "difficulty": "0x0",
-          "number": 21515063,
-          "gas_limit": 30000000,
-          "gas_used": 17303083,
-          "timestamp": 1735558967,
-          "extra_data": "0x546974616e2028746974616e6275696c6465722e78797a29",
-          "mix_hash": "0x01a0222eed85c5449615d86fbefaff1f229d6a9194f26b06fbf090921c16e80c",
-          "nonce": "0x0000000000000000",
-          "base_fee_per_gas": "0xd51e35f9",
-          "withdrawals_root": "0xbab990d946da260ba7cdafecd9ed8dae08138313d1455e1b941528f0b23dfb2a",
-          "blob_gas_used": 786432,
-          "excess_blob_gas": 76283904,
-          "parent_beacon_block_root": "0x0991ced070f670acb45f06b19248170a8c08b2ca2098241a2feb55d3d2ecd7bc"
-        }
-      }
-    ], [
-      {
-        "block_no": {
-          "Number": 21515064
-        }
-      },
-      {
-        "Eth": {
-          "parent_hash": "0xe21509c5de697a600fcf79d53414e22efaa8d018829986e6b96cfa217a5a196f",
-          "ommers_hash": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347",
-          "beneficiary": "0x0000000000000000000000000000000000000000",
-          "state_root": "0x994824cd6a75276a709b4efe42bbd1ae0b13a8da7ae438cddaf244b107ec030e",
-          "transactions_root": "0x4aad4bbc259279cc55243b9d9a4dc36c14240828ae29f8e81fd1263aaf24bfac",
-          "receipts_root": "0x63ab646d3f7eeb7f466a9baa7ea0c028ee1c6a41e63d1ba374fbc8c957435abf",
-          "logs_bloom": "0x97a155c7e5c6750c56749818b06b1fc99d6c3e984d01220184bf1813541b936006cc1799b8c3a24282f25951cc0c8da6da4580408f33fcb2962f791cd4286d08c077164938409ccb8d027939854020a3998591e18dd05e5849005c07d56c0c885f8171a48a867b66237e91befda86d0327cd8f429b09852b12a2d4fc6a5bc147087e97fbdaa45158a9c9a769fc332c524c2113d5f10590eede3f04e570d059ee866b45e25ab329317aa518c08cff9c4bc681d0bab46210748566dc4e310c385fd018d027a528049d10680bdb426f49d371863606a5204bd604179a066226696063372c2589f00828144c1785a4f0f868db58fa8840a692e001eb80105f2e7e55",
-          "difficulty": "0x0",
-          "number": 21515064,
-          "gas_limit": 30029295,
-          "gas_used": 12595115,
-          "timestamp": 1735558979,
-          "extra_data": "0x546974616e2028746974616e6275696c6465722e78797a29",
-          "mix_hash": "0x2e617a6fa120b2a0a4c26e22d806565b2056c64204834238691717899029b84b",
-          "nonce": "0x0000000000000000",
-          "base_fee_per_gas": "0xd9354fc2",
-          "withdrawals_root": "0x6ef3ae904caac578bb9338877cee0b9c8f1c0b7011a1886d16608bb9fa92a55a",
-          "blob_gas_used": 786432,
-          "excess_blob_gas": 76677120,
-          "parent_beacon_block_root": "0xc036ef6346e652e78fe530e971e302ce7e888625267f8a3240e4e619b0c2069f"
-        }
-      }
-    ]
-  ],
-  "proofs": [],
-  "transaction_count": [],
-  "balance": [],
-  "code": [],
-  "storage": []
-}
