+++ rust/services/chain/common/Cargo.toml
@@ -6,7 +6,9 @@ edition = "2021"
 [dependencies]
 alloy-primitives = { workspace = true }
 bincode = { workspace = true }
+block_header = { workspace = true, features = ["test-utils"] }
 block_trie = { workspace = true }
+bytemuck = { workspace = true }
 bytes = { workspace = true }
 common = { workspace = true }
 derivative = { workspace = true }
