+++ rust/services/chain/mock_server/Cargo.toml
@@ -7,17 +7,11 @@ edition = "2021"
 alloy-primitives = { workspace = true }
 async-trait = { workspace = true }
 axum-jrpc = { workspace = true }
-bincode = { workspace = true }
 block_header = { workspace = true, features = ["test-utils"] }
-block_trie = { workspace = true }
-bytemuck = { workspace = true }
-bytes = { workspace = true }
 chain_client = { workspace = true, features = ["rpc"] }
 chain_common = { workspace = true }
 common = { workspace = true }
 lazy_static = { workspace = true }
-risc0-zkvm = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 server_utils = { workspace = true }
