+++ rust/services/chain/mock_server/src/lib.rs
@@ -2,12 +2,11 @@ use alloy_primitives::{BlockNumber, ChainId};
 use async_trait::async_trait;
 use axum_jrpc::Value;
 use block_header::EvmBlockHeader;
-use block_trie::BlockTrie;
-use bytes::Bytes;
-use chain_common::{ChainProof, GetChainProof, GetSyncStatus, RpcChainProof, SyncStatus};
-use common::{GuestElf, Hashable, Method};
+use chain_common::{
+    fake_proof_result, ChainProof, GetChainProof, GetSyncStatus, RpcChainProof, SyncStatus,
+};
+use common::{GuestElf, Method};
 use lazy_static::lazy_static;
-use risc0_zkvm::{serde::to_vec, sha::Digest, FakeReceipt, Receipt, ReceiptClaim};
 use serde_json::json;
 use server_utils::rpc::mock::{MockBuilder, Server as RpcServerMock};
 
@@ -17,28 +16,6 @@ lazy_static! {
 }
 
 const GUEST_ELF: GuestElf = GuestElf::default();
-
-pub fn fake_proof_result(
-    guest_id: Digest,
-    block_headers: impl IntoIterator<Item = Box<dyn EvmBlockHeader>>,
-) -> RpcChainProof {
-    let mut block_trie = BlockTrie::default();
-    for header in block_headers {
-        block_trie
-            .insert_unchecked(header.number(), &header.hash_slow())
-            .expect("insert block failed");
-    }
-    let root_hash = block_trie.hash_slow();
-    let proof_output = to_vec(&(root_hash, guest_id)).unwrap();
-    let journal: Vec<u8> = bytemuck::cast_slice(&proof_output).into();
-    let inner: FakeReceipt<ReceiptClaim> =
-        FakeReceipt::<ReceiptClaim>::new(ReceiptClaim::ok(guest_id, journal.clone()));
-    let receipt = Receipt::new(risc0_zkvm::InnerReceipt::Fake(inner), journal);
-    let encoded_proof = bincode::serialize(&receipt).unwrap().into();
-    let nodes: Vec<Bytes> = block_trie.into_iter().collect();
-
-    RpcChainProof::new(encoded_proof, nodes)
-}
 
 pub struct ChainProofServerMock {
     mock_server: RpcServerMock,
